Chore: abstract validation handlers from `authorFormHandler.js` to separate modules within `/input-handlers/`
